+++ index.js
@@ -1,17 +1,24 @@
-const Json = require('./json.js');
-const Map = require('./map.js');
+const Json = require('./src/services/json');
+const Map = require('./src/services/map');
 
-const set = require('./src/handlers/set/set');
-const setAll = require('./src/handlers/set/setAll');
-const get = require('./src/handlers/get/get');
-const getAll = require('./src/handlers/get/getAll');
+const set = require('./src/services/set');
+const setAll = require('./src/services/setAll');
+const find = require('./src/services/find');
+const findAll = require('./src/services/findAll');
+const get = require('./src/services/get');
+const getAll = require('./src/services/getAll');
+const getPath = require('./src/services/getPath');
+const getPaths = require('./src/services/getPaths');
+const chop = require('./src/services/chop');
+const unlink = require('./src/handlers/basic/unlink');
 const makeObject = require('./src/handlers/make/makeObject');
 const makeArray = require('./src/handlers/make/makeArray');
 const makeJson = require('./src/handlers/make/makeJson');
 const parse = require('./src/handlers/basic/parse');
 const stringify = require('./src/handlers/basic/stringify');
-const chop = require('./src/handlers/basic/chop');
-const unlink = require('./src/handlers/basic/unlink');
+
+const resolve = require('./src/handlers/get/resolve');
+const resolveAll = require('./src/handlers/get/resolveAll');
 
 module.exports.Map = Map;
 module.exports.Json = Json;
@@ -20,6 +27,10 @@ module.exports.set = set;
 module.exports.setAll = setAll;
 module.exports.get = get;
 module.exports.getAll = getAll;
+module.exports.find = find;
+module.exports.findAll = findAll;
+module.exports.getPath = getPath;
+module.exports.getPaths = getPaths;
 module.exports.makeObject = makeObject;
 module.exports.makeArray = makeArray;
 module.exports.makeJson = makeJson;
@@ -27,3 +38,5 @@ module.exports.parse = parse;
 module.exports.stringify = stringify;
 module.exports.chop = chop;
 module.exports.unlink = unlink;
+
+module.exports.helpers = { resolve, resolveAll };