+++ src/handlers/get/src/simpleGet.js
@@ -1,9 +1,9 @@
 const validateOutput = require('./validateOutput');
 
 /**
- * Retreives single value from objects specified path from a simple format
+ * Retrieves single value from objects specified path from a simple format
  * (array representation, without queries)
- * @param {Object} obj - object/array from which value should be retreived.
+ * @param {Object} obj - object/array from which value should be retrieved.
  * @param {Array} path - array representation of path to set.
  * @returns {any} returns value found at specified path, in case that multiple logical checks
  * satisfy the first element will be returned