+++ src/helpers/pathElements/getKeys/getAllKeysFromObject.js
@@ -1,4 +1,5 @@
 const checkForError = require('./src/checkForError');
+const isObject = require('./src/isObject');
 
 /**
  * Get all keys of object, with string key
@@ -8,7 +9,7 @@ const checkForError = require('./src/checkForError');
  */
 const getAllKeysFromObject = (object, settings) => {
   const toReturn = [];
-  if (typeof object === 'object') {
+  if (isObject(object)) {
     Object.keys(object).forEach((element) => {
       toReturn.push({ string: element });
     });